use xtest (jsonrpc2_test) package to allow us to test subpackages that depend on us (in a future change)
