+++ jsonrpc2_test.go
@@ -1,4 +1,4 @@
-package jsonrpc2
+package jsonrpc2_test
 
 import (
 	"bytes"
@@ -7,15 +7,16 @@ import (
 	"fmt"
 	"io"
 	"net"
-	"reflect"
 	"strings"
 	"sync"
 	"testing"
 	"time"
+
+	"github.com/sourcegraph/jsonrpc2"
 )
 
 func TestRequest_MarshalJSON_jsonrpc(t *testing.T) {
-	b, err := json.Marshal(&Request{})
+	b, err := json.Marshal(&jsonrpc2.Request{})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -25,7 +26,7 @@ func TestRequest_MarshalJSON_jsonrpc(t *testing.T) {
 }
 
 func TestResponse_MarshalJSON_jsonrpc(t *testing.T) {
-	b, err := json.Marshal(&Response{})
+	b, err := json.Marshal(&jsonrpc2.Response{})
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -35,12 +36,12 @@ func TestResponse_MarshalJSON_jsonrpc(t *testing.T) {
 }
 
 func TestResponseMarshalJSON_Notif(t *testing.T) {
-	tests := map[*Request]bool{
-		&Request{ID: ID{Num: 0}}:                   true,
-		&Request{ID: ID{Num: 1}}:                   true,
-		&Request{ID: ID{Str: "", IsString: true}}:  true,
-		&Request{ID: ID{Str: "a", IsString: true}}: true,
-		&Request{Notif: true}:                      false,
+	tests := map[*jsonrpc2.Request]bool{
+		&jsonrpc2.Request{ID: jsonrpc2.ID{Num: 0}}:                   true,
+		&jsonrpc2.Request{ID: jsonrpc2.ID{Num: 1}}:                   true,
+		&jsonrpc2.Request{ID: jsonrpc2.ID{Str: "", IsString: true}}:  true,
+		&jsonrpc2.Request{ID: jsonrpc2.ID{Str: "a", IsString: true}}: true,
+		&jsonrpc2.Request{Notif: true}:                               false,
 	}
 	for r, wantIDKey := range tests {
 		b, err := json.Marshal(r)
@@ -63,7 +64,7 @@ func TestResponseUnmarshalJSON_Notif(t *testing.T) {
 		`{"method":"f"}`:          true,
 	}
 	for s, want := range tests {
-		var r Request
+		var r jsonrpc2.Request
 		if err := json.Unmarshal([]byte(s), &r); err != nil {
 			t.Fatal(err)
 		}
@@ -76,7 +77,7 @@ func TestResponseUnmarshalJSON_Notif(t *testing.T) {
 // testHandlerA is the "server" handler.
 type testHandlerA struct{ t *testing.T }
 
-func (h *testHandlerA) Handle(ctx context.Context, conn *Conn, req *Request) {
+func (h *testHandlerA) Handle(ctx context.Context, conn *jsonrpc2.Conn, req *jsonrpc2.Request) {
 	if req.Notif {
 		return // notification
 	}
@@ -96,7 +97,7 @@ type testHandlerB struct {
 	got []string
 }
 
-func (h *testHandlerB) Handle(ctx context.Context, conn *Conn, req *Request) {
+func (h *testHandlerB) Handle(ctx context.Context, conn *jsonrpc2.Conn, req *jsonrpc2.Request) {
 	if req.Notif {
 		h.mu.Lock()
 		defer h.mu.Unlock()
@@ -140,16 +141,16 @@ func TestClientServer(t *testing.T) {
 		if err != nil {
 			t.Fatal("Dial:", err)
 		}
-		testClientServer(ctx, t, NewStreamTransport(conn, VarintObjectCodec{}))
+		testClientServer(ctx, t, jsonrpc2.NewStreamTransport(conn, jsonrpc2.VarintObjectCodec{}))
 
 		lis.Close()
 		<-done // ensure Serve's error return (if any) is caught by this test
 	})
 }
 
-func testClientServer(ctx context.Context, t *testing.T, transport Transport) {
+func testClientServer(ctx context.Context, t *testing.T, transport jsonrpc2.Transport) {
 	hb := testHandlerB{t: t}
-	cc := NewConn(ctx, transport, &hb)
+	cc := jsonrpc2.NewConn(ctx, transport, &hb)
 	defer func() {
 		if err := cc.Close(); err != nil {
 			t.Fatal(err)
@@ -177,7 +178,7 @@ func testClientServer(ctx context.Context, t *testing.T, transport Transport) {
 
 type noopHandler struct{}
 
-func (noopHandler) Handle(ctx context.Context, conn *Conn, req *Request) {}
+func (noopHandler) Handle(ctx context.Context, conn *jsonrpc2.Conn, req *jsonrpc2.Request) {}
 
 type readWriteCloser struct {
 	read, write func(p []byte) (n int, err error)
@@ -198,7 +199,7 @@ func eof(p []byte) (n int, err error) {
 }
 
 func TestConn_DisconnectNotify_EOF(t *testing.T) {
-	c := NewConn(context.Background(), NewStreamTransport(&readWriteCloser{eof, eof}, VarintObjectCodec{}), nil)
+	c := jsonrpc2.NewConn(context.Background(), jsonrpc2.NewStreamTransport(&readWriteCloser{eof, eof}, jsonrpc2.VarintObjectCodec{}), nil)
 	select {
 	case <-c.DisconnectNotify():
 	case <-time.After(200 * time.Millisecond):
@@ -207,7 +208,7 @@ func TestConn_DisconnectNotify_EOF(t *testing.T) {
 }
 
 func TestConn_DisconnectNotify_Close(t *testing.T) {
-	c := NewConn(context.Background(), NewStreamTransport(&readWriteCloser{eof, eof}, VarintObjectCodec{}), nil)
+	c := jsonrpc2.NewConn(context.Background(), jsonrpc2.NewStreamTransport(&readWriteCloser{eof, eof}, jsonrpc2.VarintObjectCodec{}), nil)
 	if err := c.Close(); err != nil {
 		t.Error(err)
 	}
@@ -220,9 +221,9 @@ func TestConn_DisconnectNotify_Close(t *testing.T) {
 
 func TestConn_DisconnectNotify_Close_async(t *testing.T) {
 	done := make(chan struct{})
-	c := NewConn(context.Background(), NewStreamTransport(&readWriteCloser{eof, eof}, VarintObjectCodec{}), nil)
+	c := jsonrpc2.NewConn(context.Background(), jsonrpc2.NewStreamTransport(&readWriteCloser{eof, eof}, jsonrpc2.VarintObjectCodec{}), nil)
 	go func() {
-		if err := c.Close(); err != nil && err != ErrClosed {
+		if err := c.Close(); err != nil && err != jsonrpc2.ErrClosed {
 			t.Error(err)
 		}
 		close(done)
@@ -236,15 +237,15 @@ func TestConn_DisconnectNotify_Close_async(t *testing.T) {
 }
 
 func TestConn_Close_waitingForResponse(t *testing.T) {
-	c := NewConn(context.Background(), NewStreamTransport(&readWriteCloser{eof, eof}, VarintObjectCodec{}), noopHandler{})
+	c := jsonrpc2.NewConn(context.Background(), jsonrpc2.NewStreamTransport(&readWriteCloser{eof, eof}, jsonrpc2.VarintObjectCodec{}), noopHandler{})
 	done := make(chan struct{})
 	go func() {
-		if err := c.Call(context.Background(), "m", nil, nil); err != ErrClosed {
-			t.Errorf("got error %v, want %v", err, ErrClosed)
+		if err := c.Call(context.Background(), "m", nil, nil); err != jsonrpc2.ErrClosed {
+			t.Errorf("got error %v, want %v", err, jsonrpc2.ErrClosed)
 		}
 		close(done)
 	}()
-	if err := c.Close(); err != nil && err != ErrClosed {
+	if err := c.Close(); err != nil && err != jsonrpc2.ErrClosed {
 		t.Error(err)
 	}
 	select {
@@ -255,64 +256,12 @@ func TestConn_Close_waitingForResponse(t *testing.T) {
 	<-done
 }
 
-func TestAnyMessage(t *testing.T) {
-	tests := map[string]struct {
-		request, response bool
-	}{
-		// Single messages
-		`{}`:                                   {},
-		`{"foo":"bar"}`:                        {},
-		`{"method":"m"}`:                       {request: true},
-		`{"result":123}`:                       {response: true},
-		`{"error":{"code":456,"message":"m"}}`: {response: true},
-	}
-	for s, want := range tests {
-		var m anyMessage
-		json.Unmarshal([]byte(s), &m)
-		if (m.request != nil) != want.request {
-			t.Errorf("%s: got request %v, want %v", s, m.request != nil, want.request)
-		}
-		if (m.response != nil) != want.response {
-			t.Errorf("%s: got response %v, want %v", s, m.response != nil, want.response)
-		}
-	}
-}
-
-func TestMessageCodec(t *testing.T) {
-	tests := []struct {
-		v, vempty interface{}
-	}{
-		{
-			v:      &Request{ID: ID{Num: 123}},
-			vempty: &Request{ID: ID{Num: 123}},
-		},
-		{
-			v:      &Response{ID: ID{Num: 123}},
-			vempty: &Response{ID: ID{Num: 123}},
-		},
-	}
-	for _, test := range tests {
-		b, err := json.Marshal(test.v)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if err := json.Unmarshal(b, test.vempty); err != nil {
-			t.Fatal(err)
-		}
-
-		if !reflect.DeepEqual(test.vempty, test.v) {
-			t.Errorf("got %+v, want %+v", test.vempty, test.v)
-		}
-	}
-}
-
-func serve(ctx context.Context, lis net.Listener, h Handler, opt ...ConnOpt) error {
+func serve(ctx context.Context, lis net.Listener, h jsonrpc2.Handler, opt ...jsonrpc2.ConnOpt) error {
 	for {
 		conn, err := lis.Accept()
 		if err != nil {
 			return err
 		}
-		NewConn(ctx, NewStreamTransport(conn, VarintObjectCodec{}), h, opt...)
+		jsonrpc2.NewConn(ctx, jsonrpc2.NewStreamTransport(conn, jsonrpc2.VarintObjectCodec{}), h, opt...)
 	}
 }
